requirements.txt
streamlit
streamlit-quill
pandas
plotly
psycopg2
SQLAlchemy
reportlab
pillow
fastapi
uvicorn
pytest
pytest-cov
dotenv
pdfkit
matplotlib
PyPDF2
